feat: redesign search creation as 3-step wizard with shared form components
Extracted SearchFormFields.tsx with shared form components (VehicleFields,
BudgetFields, AdvancedFields), types (SearchFormData), and utilities
(formToPayload, formToUpdatePayload, isFormValid, normalizeSellerFilter).

NewSearchPage now uses a 3-step wizard:
1. Vehicle — source, manufacturer, model, search name
2. Budget — year range, price range, mileage, ownership
3. Filters — gearbox, engine, seller, keywords, photo/price toggles

Step indicator with clickable numbered badges shows progress. Presets
remain on step 1 only. Navigation with Previous/Next/Create buttons
in sticky bottom bar on mobile.

EditSearchPage reduced from 422 to 138 lines by importing shared
BudgetFields and AdvancedFields — no more form code duplication.

closes #773
closes #785
closes #790
