+++ web/src/pages/EditSearchPage.tsx
@@ -2,57 +2,19 @@ import { useState, useEffect, useRef } from "react";
 import { useNavigate, useParams, Link } from "react-router";
 import { Save, Loader2, ArrowRight } from "lucide-react";
 import { useSearch, useUpdateSearch } from "@/hooks/useSearches";
-import { formatPrice } from "@/lib/utils";
 import { Button } from "@/components/ui/Button";
-import { ChipButton } from "@/components/ui/ChipButton";
-import { Input } from "@/components/ui/Input";
-import { RangeSlider } from "@/components/ui/RangeSlider";
-import { FormField } from "@/components/ui/FormField";
 import { Skeleton } from "@/components/ui/Skeleton";
 import { EmptyState } from "@/components/ui/EmptyState";
 import { useToast } from "@/components/ui/Toast";
-
-const HAND_OPTIONS = [0, 1, 2, 3, 4];
-
-const SELLER_FILTER_OPTIONS: {
-  value: "any" | "private" | "commercial";
-  label: string;
-}[] = [
-  { value: "any", label: "הכל" },
-  { value: "private", label: "מוכר פרטי" },
-  { value: "commercial", label: "מוסך / סוכנות" },
-];
-
-const GEARBOX_OPTIONS = [
-  { value: "", label: "הכל" },
-  { value: "אוטומט", label: "אוטומט" },
-  { value: "ידני", label: "ידני" },
-];
-
-function normalizeSellerFilter(v: string | undefined): "any" | "private" | "commercial" {
-  const s = (v ?? "any").toLowerCase().trim();
-  if (s === "private") return "private";
-  if (s === "commercial" || s === "dealer" || s === "dealership") {
-    return "commercial";
-  }
-  return "any";
-}
-
-function formatKmLabel(value: number): string {
-  if (value === 0) return "ללא הגבלה";
-  return `${value.toLocaleString("he-IL")} ק"מ`;
-}
-
-function SectionCard({ title, children }: { title: string; children: React.ReactNode }) {
-  return (
-    <div className="bg-card border border-border rounded-2xl overflow-hidden">
-      <div className="px-6 py-3.5 border-b border-border bg-secondary/30">
-        <h2 className="text-xs font-semibold text-muted-foreground uppercase tracking-wider">{title}</h2>
-      </div>
-      <div className="p-6 space-y-4">{children}</div>
-    </div>
-  );
-}
+import {
+  type SearchFormData,
+  defaultFormData,
+  normalizeSellerFilter,
+  formToUpdatePayload,
+  isFormValid,
+  BudgetFields,
+  AdvancedFields,
+} from "@/components/SearchFormFields";
 
 export function EditSearchPage() {
   const navigate = useNavigate();
@@ -62,28 +24,14 @@ export function EditSearchPage() {
   const updateSearch = useUpdateSearch();
   const { toast } = useToast();
   const [error, setError] = useState<string | null>(null);
-
-  const [form, setForm] = useState({
-    yearMin: 0,
-    yearMax: 0,
-    priceMin: 0,
-    priceMax: 0,
-    engineMinCC: 0,
-    maxKm: 0,
-    maxHand: 0,
-    keywords: "",
-    excludeKeys: "",
-    sellerFilter: "any" as "any" | "private" | "commercial",
-    gearBox: "",
-    priceOnly: false,
-    photoOnly: false,
-  });
+  const [form, setForm] = useState<SearchFormData>(defaultFormData);
 
   const initializedSearchIdRef = useRef<number | null>(null);
 
   useEffect(() => {
     if (search && initializedSearchIdRef.current !== search.id) {
-      setForm({
+      setForm((prev) => ({
+        ...prev,
         yearMin: search.year_min,
         yearMax: search.year_max,
         priceMin: search.price_min ?? 0,
@@ -97,43 +45,21 @@ export function EditSearchPage() {
         gearBox: search.gear_box ?? "",
         priceOnly: search.price_only ?? false,
         photoOnly: search.photo_only ?? false,
-      });
+      }));
       initializedSearchIdRef.current = search.id;
     }
   }, [search]);
 
-  const set = <K extends keyof typeof form>(key: K, val: (typeof form)[K]) =>
+  const set = <K extends keyof SearchFormData>(key: K, val: SearchFormData[K]) =>
     setForm((prev) => ({ ...prev, [key]: val }));
 
-  const validYear = (y: number) => y === 0 || y >= 1990;
-  const canSubmit =
-    validYear(form.yearMin) &&
-    validYear(form.yearMax) &&
-    (form.yearMin === 0 || form.yearMax === 0 || form.yearMin <= form.yearMax) &&
-    !updateSearch.isPending;
+  const canSubmit = isFormValid(form) && !updateSearch.isPending;
 
   function handleFormSubmit(e: React.FormEvent<HTMLFormElement>) {
     e.preventDefault();
     setError(null);
     updateSearch.mutate(
-      {
-        id: searchId,
-        data: {
-          year_min: form.yearMin,
-          year_max: form.yearMax,
-          price_min: form.priceMin || undefined,
-          price_max: form.priceMax,
-          engine_min_cc: form.engineMinCC || undefined,
-          max_km: form.maxKm,
-          max_hand: form.maxHand,
-          keywords: form.keywords || undefined,
-          exclude_keys: form.excludeKeys || undefined,
-          seller_filter: form.sellerFilter,
-          gear_box: form.gearBox || undefined,
-          price_only: form.priceOnly,
-          photo_only: form.photoOnly,
-        },
-      },
+      { id: searchId, data: formToUpdatePayload(form) },
       {
         onSuccess: () => {
           toast("החיפוש עודכן בהצלחה!", "success");
@@ -171,16 +97,15 @@ export function EditSearchPage() {
 
   return (
     <div className="space-y-5 pb-24 md:pb-8 dir-rtl">
-      {/* Header */}
-      <header className="flex flex-col gap-1 pb-2">
+      <header className="flex flex-col gap-1">
         <Link
           to="/dashboard"
-          className="mb-2 inline-flex items-center gap-1.5 rounded-xl bg-secondary/60 px-3 py-1.5 text-sm text-muted-foreground transition-colors hover:bg-secondary hover:text-foreground w-fit"
+          className="mb-2 inline-flex items-center gap-1.5 rounded-lg bg-secondary/60 px-3 py-1.5 text-sm text-muted-foreground transition-colors hover:bg-secondary hover:text-foreground w-fit"
         >
           <span>חזרה</span>
           <ArrowRight className="h-4 w-4 shrink-0" aria-hidden />
         </Link>
-        <h1 className="text-2xl font-semibold tracking-tight text-foreground">
+        <h1 className="text-xl font-bold tracking-tight text-foreground sm:text-2xl">
           עריכת {search.manufacturer_name} {search.model_name}
         </h1>
         <p className="text-sm text-muted-foreground">מקור: {search.source}</p>
@@ -188,7 +113,7 @@ export function EditSearchPage() {
 
       {error && (
         <div
-          className="rounded-2xl border border-destructive/20 bg-destructive/5 p-4 text-sm text-destructive"
+          className="rounded-xl border border-destructive/20 bg-destructive/5 p-4 text-sm text-destructive"
           role="alert"
         >
           {error}
@@ -196,210 +121,15 @@ export function EditSearchPage() {
       )}
 
       <form onSubmit={handleFormSubmit} className="space-y-5">
-        {/* Vehicle Details (year range) */}
-        <SectionCard title="סינון לפי רכב">
-          <div className="grid gap-4 sm:grid-cols-2">
-            <FormField
-              label="שנה מ-"
-              htmlFor="yearMin"
-              error={
-                form.yearMin > form.yearMax
-                  ? "שנה מינימלית חייבת להיות קטנה מהמקסימלית"
-                  : undefined
-              }
-            >
-              <Input
-                id="yearMin"
-                type="number"
-                value={form.yearMin}
-                onChange={(e) => set("yearMin", Number(e.target.value))}
-                min={1990}
-                max={2030}
-                error={form.yearMin > form.yearMax}
-                className="tabular-nums"
-              />
-            </FormField>
+        <BudgetFields form={form} set={set} />
+        <AdvancedFields form={form} set={set} />
 
-            <FormField label="שנה עד" htmlFor="yearMax">
-              <Input
-                id="yearMax"
-                type="number"
-                value={form.yearMax}
-                onChange={(e) => set("yearMax", Number(e.target.value))}
-                min={1990}
-                max={2030}
-                className="tabular-nums"
-              />
-            </FormField>
-          </div>
-        </SectionCard>
-
-        {/* Price & Mileage */}
-        <SectionCard title='מחיר וק"מ'>
-          <div className="grid gap-4 sm:grid-cols-2">
-            <FormField
-              label="מחיר מינימום (₪)"
-              htmlFor="priceMin"
-              hint={form.priceMin > 0 ? formatPrice(form.priceMin) : undefined}
-            >
-              <Input
-                id="priceMin"
-                type="number"
-                value={form.priceMin || ""}
-                onChange={(e) => set("priceMin", Number(e.target.value))}
-                placeholder="ללא הגבלה"
-                className="tabular-nums"
-              />
-            </FormField>
-
-            <FormField
-              label="מחיר מקסימום (₪)"
-              htmlFor="priceMax"
-              hint={form.priceMax > 0 ? formatPrice(form.priceMax) : undefined}
-            >
-              <Input
-                id="priceMax"
-                type="number"
-                value={form.priceMax || ""}
-                onChange={(e) => set("priceMax", Number(e.target.value))}
-                placeholder="ללא הגבלה"
-                className="tabular-nums"
-              />
-            </FormField>
-          </div>
-
-          <div className="grid gap-4 sm:grid-cols-2">
-            <FormField label='ק"מ מקסימלי'>
-              <RangeSlider
-                min={0}
-                max={400_000}
-                step={10_000}
-                value={form.maxKm}
-                onChange={(v) => set("maxKm", v)}
-                formatLabel={formatKmLabel}
-              />
-            </FormField>
-
-            <FormField label="יד מקסימלית">
-              <div className="flex flex-wrap gap-2">
-                {HAND_OPTIONS.map((h) => (
-                  <ChipButton
-                    key={h}
-                    selected={form.maxHand === h}
-                    onClick={() => set("maxHand", h)}
-                  >
-                    {h === 0 ? "כל היידות" : `יד ${h}`}
-                  </ChipButton>
-                ))}
-              </div>
-            </FormField>
-          </div>
-        </SectionCard>
-
-        {/* Advanced Filters */}
-        <SectionCard title="פילטרים נוספים">
-          <FormField label="תיבת הילוכים">
-            <div className="flex flex-wrap gap-2">
-              {GEARBOX_OPTIONS.map((opt) => (
-                <ChipButton
-                  key={opt.value}
-                  selected={form.gearBox === opt.value}
-                  onClick={() => set("gearBox", opt.value)}
-                >
-                  {opt.label}
-                </ChipButton>
-              ))}
-            </div>
-          </FormField>
-
-          <FormField
-            label='נפח מנוע מינימלי (סמ"ק)'
-            htmlFor="engineMinCC"
-            hint={form.engineMinCC > 0 ? `${(form.engineMinCC / 1000).toFixed(1)}L` : undefined}
-          >
-            <Input
-              id="engineMinCC"
-              type="number"
-              value={form.engineMinCC || ""}
-              onChange={(e) => set("engineMinCC", Number(e.target.value))}
-              placeholder="ללא הגבלה"
-              className="tabular-nums"
-            />
-          </FormField>
-
-          <div className="space-y-1">
-            <span className="text-sm font-medium text-foreground">סוג מוכר</span>
-            <p className="text-xs text-muted-foreground">
-              מסנן לפי מודעות ממוכר פרטי או ממוסך/סוכנות.
-            </p>
-            <div className="flex flex-wrap gap-2">
-              {SELLER_FILTER_OPTIONS.map((opt) => (
-                <ChipButton
-                  key={opt.value}
-                  selected={form.sellerFilter === opt.value}
-                  onClick={() => set("sellerFilter", opt.value)}
-                >
-                  {opt.label}
-                </ChipButton>
-              ))}
-            </div>
-          </div>
-
-          <div className="space-y-1">
-            <span className="text-sm font-medium text-foreground">סינון מודעות</span>
-            <p className="text-xs text-muted-foreground">
-              הצג רק מודעות שעומדות בתנאים הבאים.
-            </p>
-            <div className="flex flex-wrap gap-2">
-              <ChipButton
-                selected={form.priceOnly}
-                onClick={() => set("priceOnly", !form.priceOnly)}
-              >
-                עם מחיר בלבד
-              </ChipButton>
-              <ChipButton
-                selected={form.photoOnly}
-                onClick={() => set("photoOnly", !form.photoOnly)}
-              >
-                עם תמונה בלבד
-              </ChipButton>
-            </div>
-          </div>
-
-          <FormField
-            label="כלול מילים"
-            htmlFor="keywords"
-            hint="הפרד מילים בפסיקים"
-          >
-            <Input
-              id="keywords"
-              value={form.keywords}
-              onChange={(e) => set("keywords", e.target.value)}
-              placeholder='לדוגמה: אוטומט, היברידי, לא פגע...'
-            />
-          </FormField>
-
-          <FormField
-            label="סנן מילים"
-            htmlFor="excludeKeys"
-            hint="מודעות שמכילות מילים אלה לא יוצגו"
-          >
-            <Input
-              id="excludeKeys"
-              value={form.excludeKeys}
-              onChange={(e) => set("excludeKeys", e.target.value)}
-              placeholder='לדוגמה: חירום, תאונה'
-            />
-          </FormField>
-        </SectionCard>
-
-        {/* Action buttons */}
-        <div className="sticky bottom-[calc(4rem+env(safe-area-inset-bottom,0px))] landscape:bottom-14 md:bottom-0 z-40 -mx-4 px-4 py-3 bg-background/90 backdrop-blur-xl border-t border-border/30 md:static md:mx-0 md:px-0 md:py-0 md:bg-transparent md:backdrop-blur-none md:border-0">
+        <div className="sticky bottom-[calc(4.5rem+env(safe-area-inset-bottom,0px))] landscape:bottom-14 md:bottom-0 z-40 -mx-4 px-4 py-3 bg-background/90 backdrop-blur-xl border-t border-border/30 md:static md:mx-0 md:px-0 md:py-0 md:bg-transparent md:backdrop-blur-none md:border-0">
           <div className="flex items-center gap-3">
             <button
               type="submit"
               disabled={!canSubmit}
-              className="flex-1 md:flex-none inline-flex items-center justify-center gap-2 bg-primary rounded-2xl px-6 py-3 text-sm font-semibold text-white shadow-lg shadow-primary/20 transition-all hover:-translate-y-px hover:shadow-xl hover:shadow-primary/30 disabled:opacity-50 disabled:cursor-not-allowed disabled:hover:translate-y-0 disabled:hover:shadow-lg"
+              className="flex-1 md:flex-none inline-flex items-center justify-center gap-2 bg-primary rounded-xl px-6 py-3 text-sm font-semibold text-white shadow-[0_4px_14px_-4px_var(--color-glow-primary)] transition-all duration-150 hover:opacity-95 disabled:opacity-50 disabled:cursor-not-allowed"
             >
               {updateSearch.isPending ? (
                 <Loader2 className="h-4 w-4 animate-spin" />
@@ -410,7 +140,7 @@ export function EditSearchPage() {
             </button>
             <Link
               to="/dashboard"
-              className="inline-flex items-center justify-center rounded-2xl border border-border px-6 py-3 text-sm font-medium text-muted-foreground transition-colors hover:bg-secondary md:flex-none"
+              className="inline-flex items-center justify-center rounded-xl border border-border px-6 py-3 text-sm font-medium text-muted-foreground transition-colors hover:bg-secondary md:flex-none"
             >
               ביטול
             </Link>
